requirements.txt
streamlit>=1.35,<2
pandas>=2.0,<3
numpy>=1.24,<3
matplotlib>=3.7,<4
plotly>=5.18,<7
python-dotenv>=1.0,<2
groq>=0.9,<2
streamlit-plotly-events>=0.0.6,<1
